chore: restructure screens, linting
