+++ frontend/app/_layout.tsx
@@ -1,13 +1,12 @@
 import { DarkTheme, DefaultTheme, ThemeProvider } from "@react-navigation/native";
 import { useFonts } from "expo-font";
-import { router, Stack } from "expo-router";
+import { Stack } from "expo-router";
 import * as SplashScreen from "expo-splash-screen";
 import { StatusBar } from "expo-status-bar";
 import { useEffect } from "react";
 import "react-native-reanimated";
 
 import { useColorScheme } from "react-native";
-import { Ionicons } from "@expo/vector-icons";
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync();
@@ -41,24 +40,6 @@ export default function RootLayout() {
                 }}>
                 <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
                 <Stack.Screen name="+not-found" />
-                <Stack.Screen
-                    name={"feed"}
-                    options={{
-                        title: "PlateMate",
-                        headerLeft: () => (
-                            <Ionicons name={"chevron-back"} size={24} color={"#20c1e6"} onPress={() => router.back()} />
-                        ),
-                    }}
-                />
-                <Stack.Screen
-                    name={"Dev1"}
-                    options={{
-                        title: "Development Environment 1",
-                        headerLeft: () => (
-                            <Ionicons name={"chevron-back"} size={24} color={"#20c1e6"} onPress={() => router.back()} />
-                        ),
-                    }}
-                />
             </Stack>
             <StatusBar style="auto" />
         </ThemeProvider>
